requirements.txt
azure-core==1.35.1
azure-functions==1.24.0
azure-storage-blob==12.26.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.2
idna==3.11
isodate==0.7.2
MarkupSafe==3.0.3
pycparser==2.23
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
